refactor: get_checkpoint
